requirements.txt
flask
pillow
requests
python-dotenv